Expermental changes
I'v made some changes to code to keep mostly consistent.

There is currenly a bug that causes some stuff to not render correctly, i'v fix this in LMC it seems to do with models that include particles, like the witch eye particles.

This is not true for the hats plugin and the thirdperson plugins by silvers doing the same thing while LMC model is witch, it can cause issue with rendering.

Ill make somechanges to silvers plugins and change the hacky thirdperson to something more easer for other plugins to play with.

Thanks for pull request, the menu is working great thanks :D

Ill do more testing and play with silver'ss plugins somemore.

Thanks.